fix(auth): isolate loader from clientMiddleware for split-route-modules
The build failed under v8_splitRouteModules: 'enforce' because the inline
`redirect()` in the new loader shared the `redirect` import binding with
clientMiddleware. When the splitter traces redirect's referencePaths, it
pulls every helper that uses `redirect` (routeGuardMiddleware,
buildRedirectWithReturnUrl, …) into loader's dependency set, which then
overlaps with clientMiddleware's set and blocks the chunk split.

Extract the loader body to a separate .server.ts helper. The route file
now references only the helper import, so loader's top-level-non-module
dependency set is empty and chunkability is restored.
